Rename atstccfg to t3c-generate
