+++ cache-config/t3c-generate/.gitignore
@@ -15,4 +15,4 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-atstccfg
+t3c-generate
