+++ cache-config/t3c-generate/cfgfile/all.go
@@ -33,7 +33,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/apache/trafficcontrol/cache-config/atstccfg/config"
+	"github.com/apache/trafficcontrol/cache-config/t3c-generate/config"
 	"github.com/apache/trafficcontrol/lib/go-atscfg"
 	"github.com/apache/trafficcontrol/lib/go-log"
 	"github.com/apache/trafficcontrol/lib/go-rfc"
@@ -147,7 +147,7 @@ var returnRegex = regexp.MustCompile(`\s*__RETURN__\s*`)
 
 // PreprocessConfigFile does global preprocessing on the given config file cfgFile.
 // This is mostly string replacements of __X__ directives. See the code for the full list of replacements.
-// These things were formerly done by ORT, but need to be processed by atstccfg now, because ORT no longer has the metadata necessary.
+// These things were formerly done by ORT, but need to be processed by t3c-generate now, because ORT no longer has the metadata necessary.
 func PreprocessConfigFile(server *atscfg.Server, cfgFile string) string {
 	if server.TCPPort != nil && *server.TCPPort != 80 && *server.TCPPort != 0 {
 		cfgFile = strings.Replace(cfgFile, `__SERVER_TCP_PORT__`, strconv.Itoa(*server.TCPPort), -1)
