+++ cache-config/t3c-generate/cfgfile/sslkeys.go
@@ -22,7 +22,7 @@ package cfgfile
 import (
 	"encoding/base64"
 
-	"github.com/apache/trafficcontrol/cache-config/atstccfg/config"
+	"github.com/apache/trafficcontrol/cache-config/t3c-generate/config"
 	"github.com/apache/trafficcontrol/lib/go-atscfg"
 	"github.com/apache/trafficcontrol/lib/go-log"
 	"github.com/apache/trafficcontrol/lib/go-tc"
