+++ cache-config/t3c-generate/getdata/getdata.go
@@ -28,8 +28,8 @@ import (
 	"os"
 	"strings"
 
-	"github.com/apache/trafficcontrol/cache-config/atstccfg/config"
-	"github.com/apache/trafficcontrol/cache-config/atstccfg/torequtil"
+	"github.com/apache/trafficcontrol/cache-config/t3c-generate/config"
+	"github.com/apache/trafficcontrol/cache-config/t3c-generate/torequtil"
 	"github.com/apache/trafficcontrol/lib/go-atscfg"
 	"github.com/apache/trafficcontrol/lib/go-log"
 	"github.com/apache/trafficcontrol/lib/go-tc"
