+++ cache-config/t3c-generate/plugin/plugin_test.go
@@ -22,7 +22,7 @@ package plugin
 import (
 	"testing"
 
-	"github.com/apache/trafficcontrol/cache-config/atstccfg/config"
+	"github.com/apache/trafficcontrol/cache-config/t3c-generate/config"
 	"github.com/apache/trafficcontrol/lib/go-atscfg"
 	"github.com/apache/trafficcontrol/lib/go-util"
 )
