+++ cache-config/t3c-generate/toreq/client.go
@@ -35,8 +35,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/apache/trafficcontrol/cache-config/atstccfg/toreqold"
-	"github.com/apache/trafficcontrol/cache-config/atstccfg/torequtil"
+	"github.com/apache/trafficcontrol/cache-config/t3c-generate/toreqold"
+	"github.com/apache/trafficcontrol/cache-config/t3c-generate/torequtil"
 	"github.com/apache/trafficcontrol/lib/go-log"
 	toclient "github.com/apache/trafficcontrol/traffic_ops/v3-client" // TODO change to v4-client when it's stabilized
 )
