+++ cache-config/t3c-generate/toreq/clientfuncs.go
@@ -1,4 +1,4 @@
-package toreqold
+package toreq
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -26,7 +26,7 @@ import (
 	"net/url"
 	"strconv"
 
-	"github.com/apache/trafficcontrol/cache-config/atstccfg/torequtil"
+	"github.com/apache/trafficcontrol/cache-config/t3c-generate/torequtil"
 	"github.com/apache/trafficcontrol/lib/go-atscfg"
 	"github.com/apache/trafficcontrol/lib/go-log"
 	"github.com/apache/trafficcontrol/lib/go-tc"
@@ -34,6 +34,10 @@ import (
 )
 
 func (cl *TOClient) GetProfileByName(profileName string) (tc.Profile, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetProfileByName(profileName)
+	}
+
 	profile := tc.Profile{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "profile_"+profileName, &profile, func(obj interface{}) error {
@@ -58,6 +62,10 @@ func (cl *TOClient) GetProfileByName(profileName string) (tc.Profile, net.Addr, 
 }
 
 func (cl *TOClient) GetGlobalParameters() ([]tc.Parameter, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetGlobalParameters()
+	}
+
 	globalParams := []tc.Parameter{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "profile_global_parameters", &globalParams, func(obj interface{}) error {
@@ -77,6 +85,10 @@ func (cl *TOClient) GetGlobalParameters() ([]tc.Parameter, net.Addr, error) {
 }
 
 func (cl *TOClient) GetServers() ([]atscfg.Server, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetServers()
+	}
+
 	servers := []atscfg.Server{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "servers", &servers, func(obj interface{}) error {
@@ -99,6 +111,10 @@ func (cl *TOClient) GetServers() ([]atscfg.Server, net.Addr, error) {
 }
 
 func (cl *TOClient) GetServerByHostName(serverHostName string) (*atscfg.Server, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetServerByHostName(serverHostName)
+	}
+
 	server := atscfg.Server{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "server-name-"+serverHostName, &server, func(obj interface{}) error {
@@ -126,6 +142,10 @@ func (cl *TOClient) GetServerByHostName(serverHostName string) (*atscfg.Server, 
 }
 
 func (cl *TOClient) GetCacheGroups() ([]tc.CacheGroupNullable, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetCacheGroups()
+	}
+
 	cacheGroups := []tc.CacheGroupNullable{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "cachegroups", &cacheGroups, func(obj interface{}) error {
@@ -150,6 +170,10 @@ func (cl *TOClient) GetCacheGroups() ([]tc.CacheGroupNullable, net.Addr, error) 
 const DeliveryServiceServersAlwaysGetAll = true
 
 func (cl *TOClient) GetDeliveryServiceServers(dsIDs []int, serverIDs []int) ([]tc.DeliveryServiceServer, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetDeliveryServiceServers(dsIDs, serverIDs)
+	}
+
 	const sortIDsInHash = true
 	toAddr := net.Addr(nil)
 	serverIDsStr := ""
@@ -216,6 +240,10 @@ func (cl *TOClient) GetDeliveryServiceServers(dsIDs []int, serverIDs []int) ([]t
 }
 
 func (cl *TOClient) GetServerProfileParameters(profileName string) ([]tc.Parameter, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetServerProfileParameters(profileName)
+	}
+
 	serverProfileParameters := []tc.Parameter{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "profile_"+profileName+"_parameters", &serverProfileParameters, func(obj interface{}) error {
@@ -236,6 +264,10 @@ func (cl *TOClient) GetServerProfileParameters(profileName string) ([]tc.Paramet
 
 // GetCDNDeliveryServices returns the data, the Traffic Ops address, and any error.
 func (cl *TOClient) GetCDNDeliveryServices(cdnID int) ([]atscfg.DeliveryService, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetCDNDeliveryServices(cdnID)
+	}
+
 	deliveryServices := []atscfg.DeliveryService{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "cdn_"+strconv.Itoa(cdnID)+"_deliveryservices", &deliveryServices, func(obj interface{}) error {
@@ -258,6 +290,10 @@ func (cl *TOClient) GetCDNDeliveryServices(cdnID int) ([]atscfg.DeliveryService,
 
 // GetTopologies returns the data, the Traffic Ops address, and any error.
 func (cl *TOClient) GetTopologies() ([]tc.Topology, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetTopologies()
+	}
+
 	topologies := []tc.Topology{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "topologies", &topologies, func(obj interface{}) error {
@@ -277,6 +313,10 @@ func (cl *TOClient) GetTopologies() ([]tc.Topology, net.Addr, error) {
 }
 
 func (cl *TOClient) GetConfigFileParameters(configFile string) ([]tc.Parameter, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetConfigFileParameters(configFile)
+	}
+
 	params := []tc.Parameter{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "config_file_"+configFile+"_parameters", &params, func(obj interface{}) error {
@@ -296,6 +336,10 @@ func (cl *TOClient) GetConfigFileParameters(configFile string) ([]tc.Parameter, 
 }
 
 func (cl *TOClient) GetCDN(cdnName tc.CDNName) (tc.CDN, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetCDN(cdnName)
+	}
+
 	cdn := tc.CDN{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "cdn_"+string(cdnName), &cdn, func(obj interface{}) error {
@@ -318,6 +362,10 @@ func (cl *TOClient) GetCDN(cdnName tc.CDNName) (tc.CDN, net.Addr, error) {
 }
 
 func (cl *TOClient) GetURLSigKeys(dsName string) (tc.URLSigKeys, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetURLSigKeys(dsName)
+	}
+
 	keys := tc.URLSigKeys{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "urlsigkeys_"+string(dsName), &keys, func(obj interface{}) error {
@@ -337,6 +385,10 @@ func (cl *TOClient) GetURLSigKeys(dsName string) (tc.URLSigKeys, net.Addr, error
 }
 
 func (cl *TOClient) GetURISigningKeys(dsName string) ([]byte, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetURISigningKeys(dsName)
+	}
+
 	keys := []byte{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "urisigningkeys_"+string(dsName), &keys, func(obj interface{}) error {
@@ -357,6 +409,10 @@ func (cl *TOClient) GetURISigningKeys(dsName string) ([]byte, net.Addr, error) {
 }
 
 func (cl *TOClient) GetParametersByName(paramName string) ([]tc.Parameter, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetParametersByName(paramName)
+	}
+
 	params := []tc.Parameter{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "parameters_name_"+paramName, &params, func(obj interface{}) error {
@@ -376,6 +432,10 @@ func (cl *TOClient) GetParametersByName(paramName string) ([]tc.Parameter, net.A
 }
 
 func (cl *TOClient) GetDeliveryServiceRegexes() ([]tc.DeliveryServiceRegexes, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetDeliveryServiceRegexes()
+	}
+
 	regexes := []tc.DeliveryServiceRegexes{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "ds_regexes", &regexes, func(obj interface{}) error {
@@ -395,6 +455,10 @@ func (cl *TOClient) GetDeliveryServiceRegexes() ([]tc.DeliveryServiceRegexes, ne
 }
 
 func (cl *TOClient) GetJobs() ([]tc.Job, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetJobs()
+	}
+
 	jobs := []tc.Job{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "jobs", &jobs, func(obj interface{}) error {
@@ -414,6 +478,10 @@ func (cl *TOClient) GetJobs() ([]tc.Job, net.Addr, error) {
 }
 
 func (cl *TOClient) GetServerCapabilitiesByID(serverIDs []int) (map[int]map[atscfg.ServerCapability]struct{}, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetServerCapabilitiesByID(serverIDs)
+	}
+
 	serverIDsStr := ""
 	if len(serverIDs) > 0 {
 		sortIDsInHash := true
@@ -452,6 +520,10 @@ func (cl *TOClient) GetServerCapabilitiesByID(serverIDs []int) (map[int]map[atsc
 }
 
 func (cl *TOClient) GetDeliveryServiceRequiredCapabilitiesByID(dsIDs []int) (map[int]map[atscfg.ServerCapability]struct{}, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetDeliveryServiceRequiredCapabilitiesByID(dsIDs)
+	}
+
 	dsIDsStr := ""
 	if len(dsIDs) > 0 {
 		sortIDsInHash := true
@@ -490,6 +562,10 @@ func (cl *TOClient) GetDeliveryServiceRequiredCapabilitiesByID(dsIDs []int) (map
 }
 
 func (cl *TOClient) GetCDNSSLKeys(cdnName tc.CDNName) ([]tc.CDNSSLKeys, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetCDNSSLKeys(cdnName)
+	}
+
 	keys := []tc.CDNSSLKeys{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "cdn_sslkeys_"+string(cdnName), &keys, func(obj interface{}) error {
@@ -509,6 +585,10 @@ func (cl *TOClient) GetCDNSSLKeys(cdnName tc.CDNName) ([]tc.CDNSSLKeys, net.Addr
 }
 
 func (cl *TOClient) GetStatuses() ([]tc.Status, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetStatuses()
+	}
+
 	statuses := []tc.Status{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "statuses", &statuses, func(obj interface{}) error {
@@ -529,6 +609,10 @@ func (cl *TOClient) GetStatuses() ([]tc.Status, net.Addr, error) {
 
 // GetServerUpdateStatus returns the data, the Traffic Ops address, and any error.
 func (cl *TOClient) GetServerUpdateStatus(cacheHostName tc.CacheName) (tc.ServerUpdateStatus, net.Addr, error) {
+	if cl.C == nil {
+		return cl.Old.GetServerUpdateStatus(cacheHostName)
+	}
+
 	status := tc.ServerUpdateStatus{}
 	toAddr := net.Addr(nil)
 	err := torequtil.GetRetry(cl.NumRetries, "server_update_status_"+string(cacheHostName), &status, func(obj interface{}) error {
