+++ cache-config/testing/ort-tests/baseline-configs/parent.config
@@ -1,4 +1,4 @@
-# DO NOT EDIT - Generated for atlanta-edge-03 by atstccfg/0.2 from https://to_server:443 ips (172.28.0.5:443) on 2021-01-13T16:46:07.704055764Z
+# DO NOT EDIT - Generated for atlanta-edge-03 by t3c-generate/0.2 from https://to_server:443 ips (172.28.0.5:443) on 2021-01-13T16:46:07.704055764Z
 # ds 'ds-top' topology 'mso-topology'
 dest_domain=origin.topology.example.net port=80 parent="atlanta-mid-16.ga.atlanta.kabletown.net:80|0.999" round_robin=consistent_hash go_direct=false qstring=ignore
 # ds '' topology ''
